+++ test/New-CustomElement.Tests.ps1
@@ -16,26 +16,26 @@ Describe "New-CustomElement" {
 	}
 
 	It 'should handle the "id" attribute' {
-		tag my-element -Id foo | Should -BeExactly '<my-element id="foo"></my-element>'
+		tag my-element -id foo | Should -BeExactly '<my-element id="foo"></my-element>'
 	}
 
 	It 'should handle the "class" attribute' {
-		tag my-element -Class btn, btn-danger | Should -BeExactly '<my-element class="btn btn-danger"></my-element>'
+		tag my-element -class btn, btn-danger | Should -BeExactly '<my-element class="btn btn-danger"></my-element>'
 	}
 
 	It 'should handle the "style" attribute' {
 		$expected = '<my-element style="font-family: &quot;Segoe UI&quot;; font-size: 1rem"></my-element>', '<my-element style="font-size: 1rem; font-family: &quot;Segoe UI&quot;"></my-element>'
-		tag my-element -Style @{ "font-family" = '"Segoe UI"'; "font-size" = "1rem" } | Should -BeIn $expected
+		tag my-element -style @{ "font-family" = '"Segoe UI"'; "font-size" = "1rem" } | Should -BeIn $expected
 	}
 
 	It "should handle custom attributes" {
 		$expected = '<my-element data-foo="&quot;bar&quot;" required />', '<my-element required data-foo="&quot;bar&quot;" />'
-		tag my-element -Attributes @{ "data-foo" = '"bar"'; disabled = $false; required = $true } -Void | Should -BeIn $expected
+		tag my-element -attributes @{ "data-foo" = '"bar"'; disabled = $false; required = $true } -Void | Should -BeIn $expected
 	}
 
 	It "should handle data attributes" {
 		$expected = '<my-element data-bs-toggle="tooltip" data-push-url></my-element>', '<my-element data-push-url data-bs-toggle="tooltip"></my-element>'
-		tag my-element -Data @{ bsToggle = "tooltip"; pushUrl = $true } | Should -BeIn $expected
+		tag my-element -data @{ bsToggle = "tooltip"; pushUrl = $true } | Should -BeIn $expected
 	}
 
 	It "should handle the inner content" {