+++ ChangeLog.md
@@ -1,9 +1,9 @@
 # Changelog
 
 ## Version [0.3.0](https://github.com/cedx/html.net/compare/v0.2.0...v0.3.0)
-- Added support for `data-*` attributes.
-- Added the `New-ImgElement` and `New-ScriptElement` cmdlets.
-- Added the `img` and `script` aliases.
+- Added support for `data-*`, `dir`, `lang` and `title` attributes.
+- Added the `New-AElement`, `New-ImgElement` and `New-ScriptElement` cmdlets.
+- Added the `a`, `img` and `script` aliases.
 
 ## Version [0.2.0](https://github.com/cedx/html.net/compare/v0.1.0...v0.2.0)
 - Added the `New-BaseElement`, `New-HtmlElement`, `New-LinkElement` and `New-MetaElement` cmdlets.